server/requirements.txt
fastapi == 0.88.0
databases == 0.6.2
openai == 0.25.0
pydantic == 1.10.2
python_dotenv == 0.14.0
uvicorn == 0.20.0
aiosqlite == 0.17.0